[FEAT] 옵션 기능 상품과 분리
