+++ src/main/java/com/sparta/quizdemo/cart/service/CartServiceImpl.java
@@ -8,9 +8,9 @@ import com.sparta.quizdemo.cart.repository.CartItemRepository;
 import com.sparta.quizdemo.cart.repository.CartRepository;
 import com.sparta.quizdemo.common.dto.ApiResponseDto;
 import com.sparta.quizdemo.common.entity.User;
-import com.sparta.quizdemo.product.entity.Option;
+import com.sparta.quizdemo.option.entity.Option;
+import com.sparta.quizdemo.option.repository.OptionRepository;
 import com.sparta.quizdemo.product.entity.Product;
-import com.sparta.quizdemo.product.repository.OptionRepository;
 import com.sparta.quizdemo.product.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
