+++ src/main/java/com/sparta/quizdemo/product/controller/ProductController.java
@@ -1,11 +1,8 @@
 package com.sparta.quizdemo.product.controller;
 
 import com.sparta.quizdemo.common.dto.ApiResponseDto;
-import com.sparta.quizdemo.product.dto.OptionRequestDto;
-import com.sparta.quizdemo.product.dto.OptionResponseDto;
 import com.sparta.quizdemo.product.dto.ProductRequestDto;
 import com.sparta.quizdemo.product.dto.ProductResponseDto;
-import com.sparta.quizdemo.product.entity.Option;
 import com.sparta.quizdemo.product.service.ProductServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -29,24 +26,10 @@ public class ProductController {
         return productService.createProduct(productRequestDto);
     }
 
-    // 옵션 생성
-    @Transactional
-    @Secured("ROLE_ADMIN")
-    @PostMapping("/product/option")
-    public ResponseEntity<OptionResponseDto> createOption(@RequestBody OptionRequestDto optionRequestDto) {
-        return productService.createOption(optionRequestDto);
-    }
-
     // 전체 상품 목록 조회
     @GetMapping("/products")
     public ResponseEntity<List<ProductResponseDto>> getProducts() {
         return productService.getProducts();
-    }
-
-    // 카테고리별 옵션 목록 조회
-    @GetMapping("/product/options/{categoryName}")
-    public ResponseEntity<List<OptionResponseDto>> getOptions(@PathVariable String categoryName) {
-        return productService.getOptions(categoryName);
     }
 
 
@@ -64,27 +47,11 @@ public class ProductController {
         return productService.updateProduct(productNo, productRequestDto);
     }
 
-    // 옵션 수정
-    @Transactional
-    @Secured("ROLE_ADMIN")
-    @PutMapping("/product/option/{optionNo}")
-    public ResponseEntity<OptionResponseDto> updateOption(@PathVariable Long optionNo, @RequestBody OptionRequestDto optionRequestDto) {
-        return productService.updateOption(optionNo, optionRequestDto);
-    }
-
     // 상품 삭제
     @Transactional
     @Secured("ROLE_ADMIN")
     @DeleteMapping("/product/{productNo}")
     public ResponseEntity<ApiResponseDto> deleteProduct(@PathVariable Long productNo) {
         return productService.deleteProduct(productNo);
-    }
-
-    // 옵션 삭제
-    @Transactional
-    @Secured("ROLE_ADMIN")
-    @DeleteMapping("/product/option/{optionNo}")
-    public ResponseEntity<ApiResponseDto> deleteOption(@PathVariable Long optionNo) {
-        return productService.deleteOption(optionNo);
     }
 }
